Edited by Matt
I added some comments and some of the functions I need. Notice that we don't just assume something like tileSize is 1, we get it from statsManager that stores all the in-game constants we might or might not change. (Then in statsManager we access those constant within the display panels Unity! That's what [SerializeField] does, it shows the variable in the panel in Unity but doesn't make it public.) Hope you approve of those changes!